feat: 공통 Stepper 추가
